backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
influxdb-client==1.38.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
reportlab==4.0.8
google-generativeai==0.8.3

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
fakeredis==2.20.0
factory-boy==3.3.0
aiosqlite==0.19.0